+++ cognite/neat/_store/_provenance.py
@@ -93,14 +93,14 @@ class Entity:
             return cls(
                 was_attributed_to=agent,
                 was_generated_by=activity,
-                id_=rules.id_,
+                id_=rules.metadata.identifier,
             )
 
         elif isinstance(rules, ReadRules | JustRules) and rules.rules is not None:
             return cls(
                 was_attributed_to=agent,
                 was_generated_by=activity,
-                id_=rules.rules.id_,
+                id_=rules.rules.metadata.identifier,
             )
         else:
             return cls(