+++ cognite/neat/_workflows/steps/data_contracts.py
@@ -10,9 +10,7 @@ from cognite.client.data_classes import (
 from cognite.client.data_classes.data_modeling import EdgeApply, NodeApply
 
 from cognite.neat._rules.models import (
-    AssetRules,
     DMSRules,
-    DomainRules,
     InformationRules,
 )
 from cognite.neat._store import NeatGraphStore
@@ -20,19 +18,13 @@ from cognite.neat._workflows.steps.step_model import DataContract
 
 
 class MultiRuleData(DataContract):
-    domain: DomainRules | None = None
     information: InformationRules | None = None
-    asset: AssetRules | None = None
     dms: DMSRules | None = None
 
     @classmethod
-    def from_rules(cls, rules: DomainRules | InformationRules | AssetRules | DMSRules):
-        if isinstance(rules, DomainRules):
-            return cls(domain=rules)
-        elif isinstance(rules, InformationRules):
+    def from_rules(cls, rules: InformationRules | DMSRules):
+        if isinstance(rules, InformationRules):
             return cls(information=rules)
-        elif isinstance(rules, AssetRules):
-            return cls(asset=rules)
         elif isinstance(rules, DMSRules):
             return cls(dms=rules)
         else: