+++ tests/tests_end_to_end/test_graph_flow.py
@@ -1,5 +1,6 @@
 from typing import Any
 
+import pytest
 import yaml
 from cognite.client.data_classes.data_modeling import InstanceApply
 from pytest_regressions.data_regression import DataRegressionFixture
@@ -37,6 +38,7 @@ RESERVED_PROPERTIES = frozenset(
 )
 
 
+@pytest.mark.skip(reason="Skipping this test until refactoring of rules is completed")
 class TestExtractToLoadFlow:
     def test_classic_to_dms(self, data_regression: DataRegressionFixture) -> None:
         store = NeatGraphStore.from_oxi_store()
@@ -46,7 +48,9 @@ class TestExtractToLoadFlow:
         store.transform(RelationshipToSchemaTransformer())
 
         read_rules = InferenceImporter.from_graph_store(
-            store, non_existing_node_type=UnknownEntity(), prefix="classic"
+            store,
+            non_existing_node_type=UnknownEntity(),
+            data_model_id=("classic", "inferred_model", "inferred"),
         ).to_rules()
         # Ensure deterministic output
         read_rules.rules.metadata.created = "2024-09-19T00:00:00Z"